+++ tests/nominal/test_MeanResponseTransformer.py
@@ -109,7 +109,7 @@ class TestInit:
         """Test that an exception is raised if weights_column is not a str."""
         with pytest.raises(
             TypeError,
-            match="MeanResponseTransformer: weights_column should be a str",
+            match="weights_column should be str or None",
         ):
             MeanResponseTransformer(weights_column=1)
 
@@ -786,7 +786,7 @@ class TestFitBinaryResponse:
 
         with pytest.raises(
             ValueError,
-            match="MeanResponseTransformer: weights column z not in X",
+            match=r"weight col \(z\) is not present in columns of data",
         ):
             x._fit_binary_response(df, df["a"], x.columns)