+++ tests/test_data.py
@@ -106,7 +106,7 @@ def create_df_6():
     """Nulls in different positions to check summing weights by col with nulls."""
     df = pd.DataFrame(
         {
-            "a": [2, 2, 2, 2, np.nan, 2, 2, 2, 3, 3],
+            "a": [2, 2, 2, 2, 0, 2, 2, 2, 3, 3],
             "b": ["a", "a", "a", "d", "e", "f", "g", np.nan, np.nan, np.nan],
             "c": ["a", "b", "c", "d", "f", "f", "f", "g", "g", np.nan],
         },
@@ -239,7 +239,7 @@ def create_weighted_imputers_test_df():
             "c": ["a", "a", np.nan, np.nan, np.nan, "f"],
             "d": [1.0, 5.0, 3.0, np.nan, np.nan, 1.0],
             "response": [0, 1, 0, 1, 1, 1],
-            "weight": [0.1, 0.1, 0.8, 0.5, 0.9, 0.8],
+            "weights_column": [0.1, 0.1, 0.8, 0.5, 0.9, 0.8],
         },
     )